Create README.md
Sample automated app aims to help the community if you lose your wallet. And another hero found it and entered the full name of the person on national card id or any unique information about your wallet that can enhance the chance to get back what you lost. The app will match and notify you with the information to connect with the hero. Also, If the lost item has a picture and the same thing in the picture is found in our collection will match the images and return back to you with information to connect with the hero!!.